move imageurl fetch back into the router.rs
